+++ react/components/AffirmSummaryPromo.tsx
@@ -7,6 +7,7 @@ import { path, prop } from 'ramda'
 
 import useScriptLoader from './hooks/useScriptLoader'
 import Settings from '../graphql/Settings.graphql'
+import { useCssHandles } from 'vtex.css-handles'
 
 interface AffirmPromoProps {
   affirmSettings: AffirmSettings
@@ -16,6 +17,8 @@ interface AffirmSettings {
   publicApiKey: string
   isLive: boolean
 }
+
+const CSS_HANDLES = ['affirmSummaryAsLowAs'] as const
 
 const AffirmSummaryPromoDiv: StorefrontFunctionComponent<AffirmPromoProps> = ({
   affirmSettings,
@@ -38,6 +41,7 @@ const AffirmSummaryPromoDiv: StorefrontFunctionComponent<AffirmPromoProps> = ({
     affirm.ui.refresh()
   }, [affirm, error, publicApiKey])
 
+  const handles = useCssHandles(CSS_HANDLES)
   const { product, isLoading } = useProductSummary()
   const commertialOffer = path<any>(
     ['sku', 'seller', 'commertialOffer'],
@@ -58,7 +62,7 @@ const AffirmSummaryPromoDiv: StorefrontFunctionComponent<AffirmPromoProps> = ({
 
   return (
     <p
-      className="affirm-as-low-as"
+      className={`affirm-as-low-as ${handles.affirmSummaryAsLowAs}`}
       data-page-type="category"
       data-amount={price * 100}
     ></p>