+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added CSS Handles to affirm components
+
+### Added
+
 - Quality Engineering Actions (SonarCloud and Lint)
 
 ## [0.0.5] - 2021-10-04